scripts: generate 3rd party reporters and services
