+++ package.json
@@ -18,7 +18,7 @@
     "setup": "run-s bootstrap build",
     "docs": "run-s docs:*",
     "docs:setup": "cd ./website && npm install && cd ..",
-    "docs:generate": "npm run clean:website && ./scripts/generateDocs.js",
+    "docs:generate": "npm run clean:website && ./scripts/docs-generation/generateDocs.js",
     "docs:build": "cd ./website && npm run build && cd ..",
     "docs:deploy": "./scripts/updateDocs.js",
     "test": "run-s test:*",
